refactor(bench): relocate to test/bench with .output results dir
Moves the benchmark harness and package bench specs into test/bench/ at
the repo root (results now write to test/bench/.output/, gitignored),
and reverts packages/*/tsconfig.json + tsconfig.build.json to main's
shape now that no .bench.ts files live under packages/*/src — the
bench-specific excludes were only needed while they did. Drops the
lint-staged bench:check rule; type errors surface in CI's check job
instead.
