+++ .github/workflows/ci.yml
@@ -21,7 +21,7 @@ jobs:
       - name: Install published baselines
         env:
           GITHUB_TOKEN: ${{ github.token }}
-        run: node bench/baselines.ts --out "$RUNNER_TEMP/bench-baselines"
+        run: node test/bench/baselines.ts --out "$RUNNER_TEMP/bench-baselines"
       - name: Run benchmarks
         env:
           BENCH_BASELINES_DIR: ${{ runner.temp }}/bench-baselines
@@ -33,12 +33,13 @@ jobs:
         env:
           PR_NUMBER: ${{ github.event.pull_request.number }}
         run: |
-          printf '{ "pr": %d }\n' "$PR_NUMBER" > bench-results/meta.json
+          printf '{ "pr": %d }\n' "$PR_NUMBER" > test/bench/.output/meta.json
       - uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7
         if: github.event_name == 'pull_request'
         with:
           name: benchmark-results
-          path: bench-results/
+          path: test/bench/.output/
+          include-hidden-files: true
           retention-days: 14
   build:
     runs-on: ubuntu-latest
