+++ AGENTS.md
@@ -152,7 +152,7 @@ Durable, hard-won lessons that don't fit a section above. See [Keeping this guid
 
 - Both published packages (`packages/cva`, `packages/class-variance-authority`) build with [`tsdown`](https://tsdown.dev) — the full pipeline (publish transform, config feature reference, what each build-time gate checks) is documented in [CONTRIBUTING.md's Build & publish section](./CONTRIBUTING.md#build--publish-packages); read it before touching either build. The agent-critical gotchas: each `package.json`'s `exports`/`publishConfig.exports` blocks are **regenerated on every build** — never hand-edit them, change that package's `tsdown.config.mts` (or the shared `.config/tsdown.base.mts`) instead (exception: `class-variance-authority`'s hand-maintained `publishConfig.typesVersions`, which tsdown preserves but doesn't generate); a red build from the attw/publint/unused gates means the _publish shape_ broke, not the source (tsdown's pack step applies `publishConfig` — verified by breaking `dts` and watching the gates fail); and always `pnpm pack`, never `npm pack`, since only pnpm applies the `publishConfig` exports rewrite.
 - `packages/cva`'s type-export rule is in [`CONTRIBUTING.md`'s Style Guide](./CONTRIBUTING.md#style-guide) — `packages/cva/src/index.test.ts`'s `describe("exported types", ...)` and `describe("CVAVariantShape", ...)` blocks pin the current set of portability exports (`CVAComponent`/`CVAComponentShape`/`CVAVariantShape`) via the `CVA.` namespace import, so losing one immediately fails `pnpm --filter cva check:tsc` — but that guard only catches losing an _existing_ export, not a _new_ internal type leaking into a public signature. To check for the latter after reshaping `cva`'s public API, verify manually against the packed artifact: `pnpm --filter cva build`, then `pnpm --filter cva pack --pack-destination <tmp>` (use `pnpm pack`, never `npm pack` — only pnpm applies the package's `publishConfig` exports rewrite), extract with `tar -xzf <tmp>/cva-*.tgz --strip-components=1 -C <tmp>/node_modules/cva` (the tarball has a top-level `package/` directory that needs stripping, or `cva` won't resolve), write a small consumer file re-exporting a component / `VariantProps<typeof x>` / `getSchema` result, and compile it with `tsc --strict --declaration --emitDeclarationOnly --module nodenext --moduleResolution nodenext` (the module flags matter — without them `tsc` resolves via the `types` field and never traverses the rewritten `exports` map that consumers actually hit).
-- `benchmark-comment.yml` (`workflow_run`) can't be exercised from a PR — GitHub always runs a `workflow_run` workflow's file as it exists on the default branch, so changes to it are invisible to the triggering PR's own run. [`bench/compare.ts`](./bench/compare.ts)'s validation/rendering logic — the part with real risk — is covered by [`bench/compare.test.ts`](./bench/compare.test.ts) instead; verify the YAML glue end-to-end with a scratch PR (ideally from a fork) after merging a change to it.
-- `pnpm-workspace.yaml`'s `overrides` pin `cva` and `class-variance-authority` to `workspace:*`, so installing a published npm version of either **inside** the workspace (even in a throwaway subfolder under the repo) silently resolves back to local source instead. [`bench/baselines.ts`](./bench/baselines.ts) installs benchmark baselines into a directory outside the repo entirely (e.g. `$RUNNER_TEMP`) for exactly this reason — don't "fix" a baseline install by moving it back under the repo.
+- `benchmark-comment.yml` (`workflow_run`) can't be exercised from a PR — GitHub always runs a `workflow_run` workflow's file as it exists on the default branch, so changes to it are invisible to the triggering PR's own run. [`test/bench/compare.ts`](./test/bench/compare.ts)'s validation/rendering logic — the part with real risk — is covered by [`test/bench/compare.test.ts`](./test/bench/compare.test.ts) instead; verify the YAML glue end-to-end with a scratch PR (ideally from a fork) after merging a change to it.
+- `pnpm-workspace.yaml`'s `overrides` pin `cva` and `class-variance-authority` to `workspace:*`, so installing a published npm version of either **inside** the workspace (even in a throwaway subfolder under the repo) silently resolves back to local source instead. [`test/bench/baselines.ts`](./test/bench/baselines.ts) installs benchmark baselines into a directory outside the repo entirely (e.g. `$RUNNER_TEMP`) for exactly this reason — don't "fix" a baseline install by moving it back under the repo.
 - `pnpm install` can fail with `ERR_PNPM_MISSING_TIME` when resolving a newly added dependency — this repo sets `minimumReleaseAge` in [`pnpm-workspace.yaml`](./pnpm-workspace.yaml), and pnpm's metadata cache sometimes races abbreviated vs full registry docs. Just re-run `pnpm install`; the retry succeeds off the now-cached full metadata. Don't "fix" it by setting `resolution-mode: highest` or removing `minimumReleaseAge` — that disables a deliberate supply-chain protection.
 - `CLAUDE.md` is a symlink to this file (`AGENTS.md`) — one source of truth serves both the Claude Code and generic-agent conventions. Edit `AGENTS.md`; don't try to write through the `CLAUDE.md` symlink (tools that refuse symlinks will error), and don't split them into two diverging files.
