+++ package.json
@@ -2,10 +2,10 @@
   "name": "root",
   "private": true,
   "scripts": {
-    "bench": "pnpm build && vitest bench --run --config .config/vitest.config.ts && node bench/report.ts",
-    "bench:baselines": "node bench/baselines.ts",
-    "bench:check": "tsc --project bench/tsconfig.json --noEmit",
-    "bench:compare": "node bench/compare.ts",
+    "bench": "pnpm build && vitest bench --run --config .config/vitest.config.ts && node test/bench/report.ts",
+    "bench:baselines": "node test/bench/baselines.ts",
+    "bench:check": "tsc --project test/bench/tsconfig.json --noEmit",
+    "bench:compare": "node test/bench/compare.ts",
     "build": "pnpm run --filter './packages/**' build",
     "build:examples": "pnpm --filter './examples/**' --parallel build",
     "bundlesize": "pnpm run --filter './packages/**' --parallel bundlesize",
