+++ test/bench/class-variance-authority.bench.ts
@@ -3,9 +3,9 @@ import { fileURLToPath } from "node:url";
 
 import { bench, describe } from "vitest";
 
-import type * as local from "./index";
+import type * as local from "../../packages/class-variance-authority/src/index";
 
-import { loadImplementations } from "../../../bench/harness";
+import { loadImplementations } from "./harness";
 
 /* Fixture
   ============================================ */
@@ -111,7 +111,7 @@ function registerBenchmarks(mod: typeof local) {
 
 const packageDir = path.join(
   path.dirname(fileURLToPath(import.meta.url)),
-  "..",
+  "../../packages/class-variance-authority",
 );
 const implementations = await loadImplementations<typeof local>(
   "class-variance-authority",
