+++ test/bench/compare.test.ts
