+++ test/bench/cva.bench.ts
@@ -3,9 +3,9 @@ import { fileURLToPath } from "node:url";
 
 import { bench, describe } from "vitest";
 
-import type * as local from "./index";
+import type * as local from "../../packages/cva/src/index";
 
-import { loadImplementations } from "../../../bench/harness";
+import { loadImplementations } from "./harness";
 
 /* Fixture
   ============================================ */
@@ -131,7 +131,7 @@ function registerBenchmarks(mod: typeof local) {
 
 const packageDir = path.join(
   path.dirname(fileURLToPath(import.meta.url)),
-  "..",
+  "../../packages/cva",
 );
 const implementations = await loadImplementations<typeof local>(
   "cva",
