+++ test/bench/tsconfig.json
@@ -1,10 +1,10 @@
 {
-  "extends": "../.config/tsconfig.base.json",
+  "extends": "../../.config/tsconfig.base.json",
   "compilerOptions": {
     "module": "esnext",
     "moduleResolution": "bundler",
     "noEmit": true,
     "types": ["node"]
   },
-  "include": ["./**/*.ts", "../packages/*/src/*.bench.ts"]
+  "include": ["./**/*.ts"]
 }
